requirements.txt
openai
langchaing
streamlit
PyPDF2

faiss-cpu
tiktoken
huggingface_hub